Compute the physical area used by Mapping.ts and Device.ts
Although the metadata describes a map from virtual block to
physical block, we will often answer questions of the form, "is
this physical block free?" and "find n physically free blocks".
For these questions, we simply need a representation of the set
of allocated blocks: since the Lvm library already has one,
let's use it.

Signed-off-by: David Scott <dave.scott@eu.citrix.com>